requirements copy.txt
# Python requirements for monitoring script
requests>=2.25.0
PyYAML>=5.4.0
